split Explorer Viewers to different script; method for loading transformer matching yolo; removed CLIP models
