+++ coralnet_toolbox/Explorer/transformer_models.py
@@ -11,9 +11,6 @@ TRANSFORMER_MODELS = {
     'DINOv2 (Base)': 'facebook/dinov2-base',
     'DINOv2 (Large)': 'facebook/dinov2-large',
     'DINOv3 ConvNext (Tiny)': 'facebook/dinov3-convnext-tiny-pretrain-lvd1689m',
-    'CLIP (Base)': 'openai/clip-vit-base-patch32',
-    'CLIP (Large)': 'openai/clip-vit-large-patch14',
-    'BioCLIP': 'imageomics/bioclip',
     'ResNet-50': 'microsoft/resnet-50',
     'ResNet-101': 'microsoft/resnet-101',
     'Swin Transformer (Tiny)': 'microsoft/swin-tiny-patch4-window7-224',
